1. Removed services and added the functionality to routers 2. Updated ingest.py to use router ingest 3. Clubed together tf artifact and data-dir as truefoundry dataloader 4. Renamed local dataloader localdir dataloader and added description to all dataloaders 5. Updated backend/requirements.txt with truefoundry[ml]==0.1.1 6. Modified all imports that use servicefoundry to use truefoundry 7. Added LOCAL env var (future use) for using local models, etc 8. Embedder provider default changed to openai and truefoundry consistent with dataloaders provider
