+++ backend/modules/embedder/__init__.py
@@ -1,14 +1,13 @@
-from servicefoundry.langchain.truefoundry_embeddings import TrueFoundryEmbeddings
+from truefoundry.langchain import TrueFoundryEmbeddings
 from langchain.embeddings.openai import OpenAIEmbeddings
 from backend.modules.embedder.embedder import register_embedder
 from backend.modules.embedder.mixbread_embedder import MixBreadEmbeddings
+from backend.settings import settings
 
-import os 
 
-OPENAI_API_KEY = os.getenv("OPENAI_API_KEY", "")
-
-if OPENAI_API_KEY:
-    register_embedder("default", OpenAIEmbeddings)
-else:
-    register_embedder("default", TrueFoundryEmbeddings)
-register_embedder("mixedbread", MixBreadEmbeddings)
+if settings.OPENAI_API_KEY:
+    register_embedder("openai", OpenAIEmbeddings)
+if settings.TFY_API_KEY:
+    register_embedder("truefoundry", TrueFoundryEmbeddings)
+if settings.LOCAL:
+    register_embedder("mixedbread", MixBreadEmbeddings)
