+++ backend/requirements.txt
@@ -14,8 +14,7 @@ orjson==3.9.15
 PyMuPDF==1.23.6
 redis==5.0.1
 beautifulsoup4==4.12.2
-mlfoundry==0.10.6
-servicefoundry==0.10.2rc4
+truefoundry[ml]==0.1.1
 markdownify==0.11.6
 gunicorn==21.2.0
 markdown-crawler==0.0.8
