+++ backend/server/routers/data_source.py
@@ -1,6 +1,9 @@
-from fastapi import APIRouter
+from fastapi import APIRouter, HTTPException
 
-from backend.server.services.data_source import DataSourceService
+from fastapi.responses import JSONResponse
+
+from backend.logger import logger
+from backend.modules.metadata_store.client import METADATA_STORE_CLIENT
 from backend.types import CreateDataSource
 
 router = APIRouter(prefix="/v1/data_source", tags=["data_source"])
@@ -9,7 +12,14 @@ router = APIRouter(prefix="/v1/data_source", tags=["data_source"])
 @router.get("/")
 def list_data_source():
     """Get data sources"""
-    return DataSourceService.list_data_sources()
+    try:
+        data_sources = METADATA_STORE_CLIENT.get_data_sources()
+        return JSONResponse(
+            content={"data_sources": [obj.dict() for obj in data_sources]}
+        )
+    except Exception as exp:
+        logger.exception(exp)
+        raise HTTPException(status_code=500, detail=str(exp))
 
 
 @router.post("/")
@@ -17,4 +27,15 @@ def add_data_source(
     data_source: CreateDataSource,
 ):
     """Create a data source for the given collection"""
-    return DataSourceService.create_data_source(data_source=data_source)
+    try:
+        created_data_source = METADATA_STORE_CLIENT.create_data_source(
+            data_source=data_source
+        )
+        return JSONResponse(
+            content={"data_source": created_data_source.dict()}, status_code=201
+        )
+    except HTTPException as exp:
+        raise exp
+    except Exception as exp:
+        logger.exception(exp)
+        raise HTTPException(status_code=500, detail=str(exp))
