--- backend/server/services/data_source.py
@@ -1,32 +1,0 @@
-from fastapi import HTTPException
-from fastapi.responses import JSONResponse
-
-from backend.logger import logger
-from backend.modules.metadata_store.client import METADATA_STORE_CLIENT
-from backend.types import CreateDataSource
-
-
-class DataSourceService:
-    def list_data_sources():
-        try:
-            data_sources = METADATA_STORE_CLIENT.get_data_sources()
-            return JSONResponse(
-                content={"data_sources": [obj.dict() for obj in data_sources]}
-            )
-        except Exception as exp:
-            logger.exception(exp)
-            raise HTTPException(status_code=500, detail=str(exp))
-
-    def create_data_source(data_source: CreateDataSource):
-        try:
-            created_data_source = METADATA_STORE_CLIENT.create_data_source(
-                data_source=data_source
-            )
-            return JSONResponse(
-                content={"data_source": created_data_source.dict()}, status_code=201
-            )
-        except HTTPException as exp:
-            raise exp
-        except Exception as exp:
-            logger.exception(exp)
-            raise HTTPException(status_code=500, detail=str(exp))
